Presenter classes inheritance added
