+++ app/src/main/java/com/pokidin/a/diary/contracts/CheckInContract.java
@@ -4,29 +4,29 @@ import com.pokidin.a.diary.common.UserData;
 
 public interface CheckInContract {
 
-    interface CheckInModel {
-        void sendCheckInUserData(UserData userData);
+//    interface CheckInModel {
+//        void sendCheckInUserData(UserData userData);
+//
+//    }
 
-    }
-
-    interface CheckInView {
-        void initView();
-
-        UserData getUserData();
-
-        void showToast(String message);
-
-        void showProgress();
-
-        void hideProgress();
-    }
-
-    interface CheckInPresenter {
-
-        void checkInBtnClicked();
-
-        void checkInUser();
-
-        void onDestroy();
-    }
+//    interface CheckInView {
+//        void initView();
+//
+//        UserData getUserData();
+//
+//        void showToast(String message);
+//
+//        void showProgress();
+//
+//        void hideProgress();
+//    }
+//
+//    interface CheckInPresenter {
+//
+//        void checkInBtnClicked();
+//
+//        void checkInUser();
+//
+//        void onDestroy();
+//    }
 }
