+++ app/src/main/java/com/pokidin/a/diary/contracts/LoginContract.java
@@ -4,27 +4,27 @@ import com.pokidin.a.diary.common.UserData;
 
 public interface LoginContract {
 
-    interface LoginModel {
-        void sendLoginUserData(UserData userData);
-    }
+//    interface LoginModel {
+//        void sendLoginUserData(UserData userData);
+//    }
 
-    interface LoginView {
-        void initView();
-
-        UserData getUserData();
-
-        void showToast(String message);
-
-        void showProgress();
-
-        void hideProgress();
-    }
-
-    interface LoginPresenter {
-        void signInBtnClicked();
-
-        void loginUser();
-
-        void onDestroy();
-    }
+//    interface LoginView {
+//        void initView();
+//
+//        UserData getUserData();
+//
+//        void showToast(String message);
+//
+//        void showProgress();
+//
+//        void hideProgress();
+//    }
+//
+//    interface LoginPresenter {
+//        void signInBtnClicked();
+//
+//        void loginUser();
+//
+//        void onDestroy();
+//    }
 }
